chore: Remove dead code and unused imports from actor and utility files
