+++ package.json
@@ -63,7 +63,6 @@
     "dedent": "^1.7.2",
     "dotenv": "^17.4.1",
     "express": "^4.21.2",
-    "mcp-client-capabilities": "^0.0.14",
     "yargs": "^17.7.2",
     "zod": "^4.1.13"
   },
