+++ src/resources/widgets.ts
@@ -5,6 +5,9 @@
  * at runtime.
  */
 
+import { existsSync } from 'node:fs';
+import path from 'node:path';
+
 import { RESOURCE_MIME_TYPE } from '@modelcontextprotocol/ext-apps';
 import type { Resource } from '@modelcontextprotocol/sdk/types.js';
 
@@ -153,15 +156,12 @@ export type AvailableWidget = WidgetConfig & {
  * @returns Map of widget URIs to their resolved state
  */
 export async function resolveAvailableWidgets(baseDir: string): Promise<Map<string, AvailableWidget>> {
-    const fs = await import('node:fs');
-    const path = await import('node:path');
-
     const resolvedWidgets = new Map<string, AvailableWidget>();
     const webDistPath = path.resolve(baseDir, '../web/dist');
 
     for (const [uri, config] of Object.entries(WIDGET_REGISTRY)) {
         const jsPath = path.resolve(webDistPath, config.jsFilename);
-        const exists = fs.existsSync(jsPath);
+        const exists = existsSync(jsPath);
 
         resolvedWidgets.set(uri, {
             ...config,
