+++ src/tools/registry.ts
@@ -83,7 +83,6 @@ export const toolCategories = {
         getDatasetList,
         getKeyValueStoreList,
     ],
-    dev: [],
 } satisfies Record<string, CategoryToolEntry[]>;
 
 /**
