+++ src/utils/tools.ts
@@ -1,9 +1,3 @@
-import { type HelperToolName } from '../const.js';
-import {
-    SKYFIRE_ENABLED_TOOLS,
-    SKYFIRE_PAY_ID_PROPERTY_DESCRIPTION,
-    SKYFIRE_TOOL_INSTRUCTIONS,
-} from '../payments/const.js';
 import type { CallDiagnostics, HelperTool, ToolBase, ToolEntry, ToolInputSchema } from '../types.js';
 import { SERVER_MODE, TOOL_TYPE } from '../types.js';
 import { fixZodSchemaRequired } from './ajv.js';
@@ -139,45 +133,4 @@ export function cloneToolEntry(toolEntry: ToolEntry): ToolEntry {
     }
 
     return cloned;
-}
-
-/** Returns true if the tool is eligible for Skyfire augmentation. */
-function isSkyfireEligible(tool: ToolEntry): boolean {
-    return (
-        tool.type === TOOL_TYPE.ACTOR ||
-        (tool.type === TOOL_TYPE.INTERNAL && SKYFIRE_ENABLED_TOOLS.has(tool.name as HelperToolName))
-    );
-}
-
-/**
- * Applies Skyfire augmentation to a tool entry.
- * Clones the tool and, if eligible, appends Skyfire instructions to the description
- * and adds a `skyfire-pay-id` property to the input schema.
- *
- * Returns the (possibly augmented) clone if the tool is eligible,
- * or the original tool reference if it is not eligible.
- * Augmentation is idempotent — calling this on an already-augmented clone is safe.
- */
-export function applySkyfireAugmentation(tool: ToolEntry): ToolEntry {
-    if (!isSkyfireEligible(tool)) return tool;
-
-    const cloned = cloneToolEntry(tool);
-
-    // Append Skyfire instructions to description (idempotent)
-    if (cloned.description && !cloned.description.includes(SKYFIRE_TOOL_INSTRUCTIONS)) {
-        cloned.description += `\n\n${SKYFIRE_TOOL_INSTRUCTIONS}`;
-    }
-
-    // Add skyfire-pay-id property to inputSchema (idempotent)
-    if (cloned.inputSchema && 'properties' in cloned.inputSchema) {
-        const props = cloned.inputSchema.properties as Record<string, unknown>;
-        if (!props['skyfire-pay-id']) {
-            props['skyfire-pay-id'] = {
-                type: 'string',
-                description: SKYFIRE_PAY_ID_PROPERTY_DESCRIPTION,
-            };
-        }
-    }
-
-    return Object.freeze(cloned);
 }
